Propagated to tests.
