+++ tools/dma_from_device.c
@@ -28,7 +28,7 @@
 
 #include "dma_utils.c"
 
-#define DEVICE_NAME_DEFAULT "/dev/xdma0_c2h_0"
+#define DEVICE_NAME_DEFAULT "/dev/zdma0_c2h_0"
 #define SIZE_DEFAULT (32)
 #define COUNT_DEFAULT (1)
